+++ Quaternion.h
@@ -75,7 +75,7 @@ class Quaternion
      * Constructs a Quaternion from four single
      * values
      */
-    Quaternion(float u, float x, float y, float z);
+    Quaternion(double u, double x, double y, double z);
 
     /*!
      * \brief Constructor
@@ -88,7 +88,7 @@ class Quaternion
     Quaternion(const Vector3 &from, const Vector3 &to);
 
     // construct a quaternion from the real part u and the imaginary vector part (x y z)
-    Quaternion(float realPart, const Vector3 vectorPart);
+    Quaternion(double realPart, const Vector3 vectorPart);
 
     /*!
      * \brief Constructor from Euler angles
@@ -118,21 +118,21 @@ class Quaternion
     void toRotMatrix(std::vector <double>& rot_matrix_3_3) const;
 
 
-    inline const float& operator() (unsigned int i) const { return data[i]; }
-    inline float& operator() (unsigned int i) { return data[i]; }
+    inline const double& operator() (unsigned int i) const { return data[i]; }
+    inline double& operator() (unsigned int i) { return data[i]; }
 
-    float norm () const;
+    double norm () const;
     Quaternion  normalized () const;
     Quaternion& normalize ();
 
     double dot(const Quaternion &other) const;
-    float getShortestAngleTo(const Quaternion &other) const;
+    double getShortestAngleTo(const Quaternion &other) const;
 
     Vector3 getImag() const;
     Vector3 getRotationAxis() const;
-    float getRotationAngle() const;
+    double getRotationAngle() const;
 
-    Quaternion lerp(const Quaternion &other, float t) const;
+    Quaternion lerp(const Quaternion &other, double t) const;
     Quaternion slerp(const Quaternion &other, double t) const;
 
     void decomposeSwingTwist(const Vector3 &direction, Quaternion* swing, Quaternion* twist) const;
@@ -141,14 +141,14 @@ class Quaternion
 
     Quaternion exp() const;
     Quaternion log() const;
-    Quaternion pow(float value) const;
+    Quaternion pow(double value) const;
 
-    Quaternion operator*(float val) const;
+    Quaternion operator*(double val) const;
     Quaternion operator+(const Quaternion &other) const;
     Quaternion operator-(const Quaternion &other) const;
 
     void operator*=(const Quaternion &other);
-    void operator/=(float x);
+    void operator/=(double x);
     Quaternion& operator= (const Quaternion& other);
     bool operator== (const Quaternion& other) const;
 
@@ -202,15 +202,15 @@ class Quaternion
      */
     Vector3 rotate(const Vector3 &v) const;
 
-    inline float& u() { return data[0]; }
-    inline float& x() { return data[1]; }
-    inline float& y() { return data[2]; }
-    inline float& z() { return data[3]; }
+    inline double& u() { return data[0]; }
+    inline double& x() { return data[1]; }
+    inline double& y() { return data[2]; }
+    inline double& z() { return data[3]; }
 
-    inline const float& u() const { return data[0]; }
-    inline const float& x() const { return data[1]; }
-    inline const float& y() const { return data[2]; }
-    inline const float& z() const { return data[3]; }
+    inline const double& u() const { return data[0]; }
+    inline const double& x() const { return data[1]; }
+    inline const double& y() const { return data[2]; }
+    inline const double& z() const { return data[3]; }
 
     std::istream& read(std::istream &s);
     std::ostream& write(std::ostream &s) const;
@@ -218,7 +218,7 @@ class Quaternion
     std::ostream& writeBinary(std::ostream &s) const;
 
   protected:
-    float data[4];
+    double data[4];
 };
 
 //! output in format u x y z
